requirements.txt
fastapi>=0.110
uvicorn>=0.29
paho-mqtt>=1.6
pyyaml>=6.0
